code cleanup

+++ common/src/configtool/config-tool-utils.h
@@ -54,13 +54,13 @@ int config_section_str_append (struct config_section_str **section_strs,
 
 void config_section_str_destroy (struct config_section_str *section_str);
 
-int8_t config_ipv4_address_string2int (pstdout_state_t pstate,
-                                       const char *src,
-                                       uint32_t *dest);
+int config_ipv4_address_string2int (pstdout_state_t pstate,
+                                    const char *src,
+                                    uint32_t *dest);
 
-int8_t config_mac_address_string2int (pstdout_state_t pstate,
-                                      const char *src,
-                                      uint64_t *dest);
+int config_mac_address_string2int (pstdout_state_t pstate,
+                                   const char *src,
+                                   uint64_t *dest);
 
 struct config_section *config_find_section (pstdout_state_t pstate,
                                             struct config_section *sections,
